Parser_OOP vs2013. Calculate s1 and s2 correctly

+++ VS2013/Parser/Parser_OOP/PeakFinder.cpp
@@ -40,7 +40,7 @@ PeakFinder::~PeakFinder()
 {
 }
 
-std::vector<double>& PeakFinder::GetPeakPosition()
+std::vector<int>& PeakFinder::GetPeakPosition()
 {
 	return peak_position;
 }
