+++ VS2013/Parser/Parser_OOP/PeakFinder.h
@@ -8,8 +8,8 @@ class PeakFinder
 public:
 	PeakFinder(std::vector<double>& yv, std::vector<double>& time, const double th);
 	~PeakFinder();
-	std::vector<double>& GetPeakPosition();
+	std::vector<int>& GetPeakPosition();
 private:
-	std::vector<double> peak_position;
+	std::vector<int> peak_position;
 };
